fix: mudando a estrutura das rotas
