+++ src/App.jsx
@@ -1,19 +1,21 @@
 import "./App.css";
-import AppRoutes from "./AppRoutes";
+import AppRoutes from "./routes/AppRoutes";
 import SnackBarComponent from "./components/atoms/snackbar/snackbar";
 import { useSelector } from "react-redux";
 import LanguageProvider from "./contexts/languageContext";
+import { AuthProvider } from "./contexts/auth";
 
 function App() {
   const { open, message } = useSelector((state) => state.snackbar);
 
   return (
-    <div className="App">
+    <AuthProvider>
       <LanguageProvider>
         <AppRoutes />
+
         <SnackBarComponent open={open} message={message} />
       </LanguageProvider>
-    </div>
+    </AuthProvider>
   );
 }
