+++ src/pages/home-page/home-page.jsx
@@ -5,13 +5,9 @@ import { Container, BoxCenter } from "./styles";
 import HeaderHome from "../../components/molecules/header-home/header-home";
 import GlobalStyles from "../../styles/global-style";
 import { TweetsListContext } from "../../contexts/tweetsHomePageContext";
-import { AuthContext } from "../../contexts/auth";
 import { i18n } from "../../translate/i18n";
-import { useTranslation } from "react-i18next";
 import Sidebar from "../../components/molecules/sidebar/sidebar";
 function HomePage() {
-  const { user } = useContext(AuthContext);
-
   const {
     tweets,
     refreshTweets,
